Feat: Implement CSRF on comments and posts in admin dashboard

+++ src/Controller/Backoffice/AdminPostController.php
@@ -6,6 +6,7 @@ use App\Model\Entity\Post;
 use App\Model\Repository\PostRepository;
 use App\Model\Repository\UserRepository;
 use App\Service\Authentication\Authentication;
+use App\Service\CSRF\Csrf;
 use App\Service\Form\PostFormValidator;
 use App\Service\Http\RedirectResponse;
 use App\Service\Http\Request;
@@ -23,8 +24,9 @@ class AdminPostController
     private UserRepository $userRepository;
     private PostFormValidator $postFormValidator;
     private Authentication $authentication;
+    private Csrf $csrf;
 
-    public function __construct(Request $request, View $view, Session $session, PostRepository $postRepository, UserRepository $userRepository, PostFormValidator $postFormValidator, Authentication $authentication)
+    public function __construct(Request $request, View $view, Session $session, PostRepository $postRepository, UserRepository $userRepository, PostFormValidator $postFormValidator, Authentication $authentication, Csrf $csrf)
     {
         $this->request = $request;
         $this->view = $view;
@@ -33,21 +35,25 @@ class AdminPostController
         $this->userRepository = $userRepository;
         $this->postFormValidator = $postFormValidator;
         $this->authentication = $authentication;
+        $this->csrf = $csrf;
     }
 
     public function displayAdminPostAction(): Response
     {
         if ($this->authentication->isAdmin() || $this->authentication->isEditor()) {
+            $this->csrf->generateToken();
+
             $posts = $this->postRepository->findAll();
 
             $data = $this->request->request()->all();
 
-            if ($this->request->getMethod() === 'POST') {
+
+            if ($this->request->getMethod() === 'POST' && $this->csrf->checkToken($data['csrfToken'])) {
 
                 if ($this->authentication->isAdmin()) {
-
                     $post = $this->postRepository->find($data['deletePost']);
                     $this->postRepository->delete($post);
+                    $this->csrf->deleteToken();
 
                     $this->session->addFlashes('success', 'Article supprimé.');
                     return new RedirectResponse('index.php?action=admin_posts', 302);
@@ -57,7 +63,8 @@ class AdminPostController
             return new Response($this->view->render([
                 'template' => '../backoffice/posts',
                 'data' => [
-                    'posts' => $posts
+                    'posts' => $posts,
+                    'csrfToken' => $this->session->get('csrfToken')
                 ]
             ]));
         }
@@ -69,11 +76,13 @@ class AdminPostController
     public function addPostAction()
     {
         if ($this->authentication->isAdmin() || $this->authentication->isEditor()) {
+            $this->csrf->generateToken();
+
             $user = $this->authentication->getAuthenticatedUser();
 
             $data = $this->request->request()->all();
 
-            if ($this->request->getMethod() === 'POST') {
+            if ($this->request->getMethod() === 'POST' && $this->csrf->checkToken($data['csrfToken'])) {
 
                 if ($this->postFormValidator->isValid($data)) {
 
@@ -86,6 +95,7 @@ class AdminPostController
                         ->setUserFk($user->getId());
 
                     $this->postRepository->create($post);
+                    $this->csrf->deleteToken();
 
                     $this->session->addFlashes('success', 'Article enregistré.');
                     return new RedirectResponse('index.php?action=admin_posts', 302);
@@ -97,7 +107,10 @@ class AdminPostController
             }
 
             return new Response($this->view->render([
-                'template' => '../backoffice/add_post'
+                'template' => '../backoffice/add_post',
+                'data' => [
+                    'csrfToken' => $this->session->get('csrfToken')
+                ]
             ]));
         }
 
@@ -107,12 +120,14 @@ class AdminPostController
     public function editPostAction(int $id)
     {
         if ($this->authentication->isAdmin() || $this->authentication->isEditor()) {
+            $this->csrf->generateToken();
+
             $user = $this->authentication->getAuthenticatedUser();
             $post = $this->postRepository->find($id);
 
             $data = $this->request->request()->all();
 
-            if ($this->request->getMethod() === 'POST') {
+            if ($this->request->getMethod() === 'POST' && $this->csrf->checkToken($data['csrfToken'])) {
 
                 if ($this->postFormValidator->isValid($data)) {
 
@@ -124,6 +139,7 @@ class AdminPostController
                         ->setUserFk($user->getId());
 
                     $this->postRepository->update($post);
+                    $this->csrf->deleteToken();
 
                     $this->session->addFlashes('success', 'Article mis à jour.');
                     return new RedirectResponse('index.php?action=admin_posts', 302);
@@ -137,7 +153,8 @@ class AdminPostController
             return new Response($this->view->render([
                 'template' => '../backoffice/edit_post',
                 'data' => [
-                    'post' => $post
+                    'post' => $post,
+                    'csrfToken' => $this->session->get('csrfToken')
                 ]
             ]));
         }
@@ -150,15 +167,18 @@ class AdminPostController
         if ($this->authentication->isAdmin() || $this->authentication->isEditor()) {
 // TODO : Ajouter confirmation de suppression
 
+            $this->csrf->generateToken();
+
             $post = $this->postRepository->find($id);
 
             $data = $this->request->request()->all();
 
-            if ($this->request->getMethod() === 'POST') {
+            if ($this->request->getMethod() === 'POST' && $this->csrf->checkToken($data['csrfToken'])) {
 
                 if ($this->authentication->isAdmin()) {
 
                     $this->postRepository->delete($post);
+                    $this->csrf->deleteToken();
 
                     $this->session->addFlashes('success', 'Article supprimé.');
                     return new RedirectResponse('index.php?action=admin_posts', 302);
